+++ ostap/utils/progress_conf.py
@@ -35,11 +35,16 @@ else                       : logger = getLogger( __name__ )
 def progress_conf ( show = True ) :
     """configuration of the progress bar"""
     if not show : return Ostap.Utils.ProgressConf ( 0 )
+
+    if isatty () : 
+        twidth = terminal_size () [ 1 ]
+        width  = twidth - 15 if 25 < twidth else 0  ## silent if is terminal is too narrow
+    else :
+        width  = 0                                  ## silent if not isatty 
     
-    tty = isatty () 
-    twidth        = terminal_size () [ 1 ] if tty else 110 
     return Ostap.Utils.ProgressConf (
-        twidth - 15 if 25 < twidth else 0     , ## silent if is terminal is too narrow
+        width                                 , 
+        ## twidth - 15 if 25 < twidth else 0  , ## silent if is terminal is too narrow
         allright ( '#'   ) if tty else '#'    , ## 'done' symbol 
         ' '                                   , ## 'not-yet' symbol 
         allright ( ' [ ' ) if tty else ' { '  , ## left